refactor: common package and testutils
