+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/irrationals/sqrt/Sqrt.kt
@@ -1,8 +1,8 @@
 package xyz.lbres.exactnumbers.irrationals.sqrt
 
-import xyz.lbres.common.createHashCode
-import xyz.lbres.common.divideBigDecimals
-import xyz.lbres.common.divideByZero
+import xyz.lbres.exactnumbers.common.createHashCode
+import xyz.lbres.exactnumbers.common.divideBigDecimals
+import xyz.lbres.exactnumbers.common.divideByZero
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumberCompanion
