+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/irrationals/sqrt/helpers.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.exactnumbers.irrationals.sqrt
 
-import xyz.lbres.common.getIntFromDecimal
+import xyz.lbres.exactnumbers.common.getIntFromDecimal
 import xyz.lbres.exactnumbers.irrationals.common.Memoize
 import xyz.lbres.kotlinutils.biginteger.ext.isNegative
 import xyz.lbres.kotlinutils.biginteger.ext.isZero
