+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/exactfraction/constructSimplifyTests.kt
@@ -1,7 +1,7 @@
 package xyz.lbres.exactnumbers.exactfraction
 
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import xyz.lbres.kotlinutils.pair.TypePair
-import xyz.lbres.testutils.assertDivByZero
 import java.math.BigInteger
 import kotlin.test.assertEquals
