+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/exactfraction/unaryNonOpTests.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.exactnumbers.exactfraction
 
-import xyz.lbres.testutils.assertDivByZero
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import java.math.RoundingMode
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
