+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/irrationals/sqrt/binaryOpTests.kt
@@ -3,9 +3,9 @@ package xyz.lbres.exactnumbers.irrationals.sqrt
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
+import xyz.lbres.exactnumbers.testutils.TestNumber
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import xyz.lbres.expressions.term.Term
-import xyz.lbres.testutils.TestNumber
-import xyz.lbres.testutils.assertDivByZero
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue
