+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/irrationals/sqrt/castingTests.kt
@@ -1,8 +1,8 @@
 package xyz.lbres.exactnumbers.irrationals.sqrt
 
-import xyz.lbres.common.CastingOverflowException
+import xyz.lbres.exactnumbers.common.CastingOverflowException
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
-import xyz.lbres.testutils.assertFailsWithMessage
+import xyz.lbres.exactnumbers.testutils.assertFailsWithMessage
 import java.math.BigInteger
 import kotlin.test.assertEquals
