+++ exact-numbers/src/test/kotlin/xyz/lbres/expressions/term/TermTest.kt
@@ -4,7 +4,7 @@ import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
-import xyz.lbres.testutils.TestNumber
+import xyz.lbres.exactnumbers.testutils.TestNumber
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
