+++ exact-numbers/src/test/kotlin/xyz/lbres/expressions/term/binaryOpsTests.kt
@@ -4,8 +4,8 @@ import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
-import xyz.lbres.testutils.TestNumber
-import xyz.lbres.testutils.assertDivByZero
+import xyz.lbres.exactnumbers.testutils.TestNumber
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import kotlin.test.assertEquals
 
 private val logNum1 = Log(ExactFraction(15, 4))
