+++ exact-numbers/src/test/kotlin/xyz/lbres/expressions/term/constuctorTests.kt
@@ -5,7 +5,7 @@ import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
-import xyz.lbres.testutils.TestNumber
+import xyz.lbres.exactnumbers.testutils.TestNumber
 import kotlin.test.assertEquals
 
 private val logWhole = Log(1000)
