+++ exact-numbers/src/test/kotlin/xyz/lbres/expressions/term/getComponentsTests.kt
@@ -5,8 +5,8 @@ import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
+import xyz.lbres.exactnumbers.testutils.TestNumber
 import xyz.lbres.kotlinutils.list.StringList
-import xyz.lbres.testutils.TestNumber
 import kotlin.test.assertEquals
 
 private val logNum1 = Log(ExactFraction(15, 4))
