+++ exact-numbers/src/test/kotlin/xyz/lbres/expressions/term/simplifyGetValuesTests.kt
@@ -5,8 +5,8 @@ import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
+import xyz.lbres.exactnumbers.testutils.TestNumber
 import xyz.lbres.kotlinutils.collection.ext.toConstMultiSet
-import xyz.lbres.testutils.TestNumber
 import java.math.BigDecimal
 import kotlin.test.assertEquals
